+++ library/src/main/java/com/twilio/video/Track.java
@@ -21,11 +21,6 @@ package com.twilio.video;
  */
 interface Track {
     /**
-     * Returns the id associated with the track.
-     */
-    String getTrackId();
-
-    /**
      * Checks if the track is enabled.
      */
     boolean isEnabled();